Fixed bugs in the "WorkingWithArrays" project.
- in the "functionForWorkingWithArrays.js" file:

1. The following functions have been renamed:
--- from "getDescendingSort" to "sortDescending";
--- from "getEvenItemsSum" to "getEvenNumbersSum";
--- from "getOneHundredItemsArray" to "getOneHundredNumbersArray";
--- from "getEvenNumbersSquaresList" to "getEvenNumbersSquaresArray".

2. The following functions have been created:
--- isNotNumericArray;
--- isArrayWithRequestedElementsCount.

3. The following function have been changed:

--- getStringSeparator
(The function for dublicating the "#" symbol has been replaced.);

--- sortDescending
(The anonymous function in the "sort" method has been changed.
The conditions for array validation have been changed to the "isNotNumericArray" function.);

--- getItemsFromBeginning
(The "filter" function has been changed to "slice".
The conditions for array validation have been changed to the "isArrayWithRequestedElementsCount" function.);

--- getItemsFromEnd
(The "filter" function has been changed to "slice".
The conditions for array validation have been changed to the "isArrayWithRequestedElementsCount" function.);

--- getEvenNumbersSum
(The conditions for array validation have been changed to the "isNotNumericArray" function.);

--- getOneHundredNumbersArray
(The "fill" and "map" functions have been replaced with a loop.);

--- getEvenNumbersSquaresArray
(The conditions for array validation have been changed to the "isNotNumericArray" function.).

4. If the arguments of the functions do not pass validation, the function returns "null".

5. The coding style has been changed.

6. The "use strict" regime has been added.